@@ -1,34 +1,42 @@
-[tool.poetry]
+[project]
 name = "netcheck"
-version = "0.5.4"
+version = "0.5.5"
 description = "Netchecks is a cloud native tool for specifying and regularly checking assertions about network conditions. Use netchecks to proactively verify whether security controls are working as intended, alerting on misconfiguration."
-authors = ["Brian Thorne <brian@hardbyte.nz>"]
 readme = "README.md"
-packages = [{include = "netcheck"}]
+authors = [
+    {name = "Brian Thorne", email = "brian@hardbyte.nz"},
+]
+requires-python = "<4.0,>=3.10"
+dependencies = [
+    "dnspython<3.0,>=2.2",
+    "requests<3.0,>=2.28",
+    "typer<1.0,>=0.9",
+    "pydantic<3.0,>=1.10",
+    "rich<14.0.0,>=10.11.0",
+    "cel-python<1.0.0,>=0.1.5",
+]
 
-[tool.poetry.dependencies]
-python = "^3.10"
-dnspython = "^2.2"
-requests = "^2.28"
-typer = "^0.9"
-pydantic = ">=1.10,<3.0"
-rich = ">=10.11.0,<14.0.0"
-cel-python = "^0.1.5"
-
-
-[tool.poetry.scripts]
-netcheck = "netcheck.cli:app"
-
-
-[tool.poetry.group.dev.dependencies]
-pytest = ">=7.2,<9.0"
-coveralls = "^3.3.1"
-pytest-cov = "^4.0.0"
-ruff = ">=0.0.241,<0.3.4"
 
 [tool.ruff]
 line-length = 120
 
+[tool.uv]
+dev-dependencies = [
+    "pytest<9.0,>=7.2",
+    "coveralls<4.0.0,>=3.3.1",
+    "pytest-cov<5.0.0,>=4.0.0",
+    "ruff<0.3.4,>=0.0.241",
+]
+
+[project.scripts]
+netcheck = "netcheck.cli:app"
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["netcheck"]
+exclude = ["operator", "scripts", "tests", "docs"]
+
+
 [build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires = ["setuptools", "wheel"]
+build-backend = "setuptools.build_meta"
