@@ -60,7 +60,7 @@ def http_request_check(
     }
 
     result_data = {
-        "startTimestamp": datetime.datetime.utcnow().isoformat(),
+        "startTimestamp": datetime.datetime.now(datetime.UTC).isoformat(),
     }
 
     output = {"spec": test_spec, "data": result_data}
@@ -83,6 +83,6 @@ def http_request_check(
         result_data["exception-type"] = e.__class__.__name__
         result_data["exception"] = str(e)
 
-    result_data["endTimestamp"] = datetime.datetime.utcnow().isoformat()
+    result_data["endTimestamp"] = datetime.datetime.now(datetime.UTC).isoformat()
 
     return output
