@@ -26,7 +26,7 @@ if __name__ == "__main__":
     operator_toml = read_toml(operator_toml_path)
     operator_chart_yaml = read_yaml(operator_chart_path)
 
-    cli_version = cli_toml["tool"]["poetry"]["version"]
+    cli_version = cli_toml["project"]["version"]
     operator_version = operator_toml["tool"]["poetry"]["version"]
     operator_chart_version = operator_chart_yaml["appVersion"]
 
